requirements.txt
mediapipe==0.10.9
pyserial
opencv-python
pandas
openpyxl